Item과 관련된 필드 수정

+++ backend/src/main/java/hanglog/expense/Expense.java
@@ -29,7 +29,7 @@ public class Expense extends BaseEntity {
     private String currency;
 
     @Column(nullable = false)
-    private Integer amount;
+    private Double amount;
 
     @ManyToOne(fetch = LAZY, cascade = CascadeType.PERSIST)
     @JoinColumn(name = "category_id", nullable = false)
@@ -38,7 +38,7 @@ public class Expense extends BaseEntity {
     public Expense(
             final Long id,
             final String currency,
-            final Integer amount,
+            final Double amount,
             final Category category
     ) {
         this.id = id;
@@ -49,7 +49,7 @@ public class Expense extends BaseEntity {
 
     public Expense(
             final String currency,
-            final Integer amount,
+            final Double amount,
             final Category category
     ) {
         this(null, currency, amount, category);
